Break out admin action button group into sub component on apps page, part 0 of breaking down the group tables into accordions
